Move liveness analysis to analysis directory
